renaming package res and log, preparations for result and inputfile update

+++ src/main/java/de/unileipzig/irpact/core/spatial/SpatialInformation.java
@@ -24,13 +24,13 @@ public interface SpatialInformation extends ChecksumComparable {
 
     boolean hasId();
 
-    boolean isId(int id);
+    boolean isId(long id);
 
     default boolean hasSameId(SpatialInformation other) {
         return getId() == other.getId();
     }
 
-    int getId();
+    long getId();
 
     AttributeAccess getAttributeAccess();
